arxiv_messaging/pyproject.toml
[tool.poetry]
name = "arxiv-messaging"
version = "0.1.0"
description = "ArXiv messaging library for event types and messaging utilities"
authors = ["ArXiv Team <team@arxiv.org>"]
readme = "README.md"
packages = [{include = "arxiv_messaging"}]

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.31.1"
google-auth = "^2.0.0"
requests = "^2.32.3"
firebase-admin = "^7.1.0"
ruamel-yaml = "^0.18.6"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
arxiv-manage-subscribers = "arxiv_messaging.manage_subscribers:main"
arxiv-send-test-message = "arxiv_messaging.send_test_message:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"